+++ src/index.ts
@@ -1,8 +1,11 @@
 import { ModuleRegistrationFunction } from '@causa/workspace';
+import { registerFunctions } from './functions/index.js';
 
 export * from './configurations/index.js';
 export * from './services/index.js';
 
-const registerModule: ModuleRegistrationFunction = async () => {};
+const registerModule: ModuleRegistrationFunction = async (context) => {
+  registerFunctions(context);
+};
 
 export default registerModule;