+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 Features:
 
 - Implement the `TerraformService`.
+- Implement the `InfrastructureDeployForTerraform` and `InfrastructurePrepareForTerraform` functions.